% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlPictographs.R
\name{graphic}
\alias{graphic}
\title{rhtmlPictographs HTML Widget}
\source{
https://github.com/Displayr/rhtmlPictographs
}
\usage{
graphic(settingsJsonString = "{}")
}
\description{
A HTMLWidget that can generate single image graphics, mutli image graphics, or a table of single/multi image graphics. See scenarios and examples below for different use cases.

The usage documentation is maintained as a self hosted tutorial. For instructions on how to run and view this tutorial see the project readme : https://github.com/Displayr/rhtmlPictographs
}
\examples{

red circle
rhtmlPictographs::graphic('circle:red')

single horizontally cropped image - minimal settings
rhtmlPictographs::graphic('{"variableImage":"circle:horizontal:blue","percentage":"0.4","width":400,"height":400}')

for a list of up to date examples see here : https://github.com/Displayr/rhtmlPictographs/tree/master/examples

}
\author{
Kyle Zeeuwen <kyle.zeeuwen@gmail.com>
}
